# backend/docs/swagger.yaml
definitions:
...
  models.RecipeIngredient:
    properties:
      ingredient_id:
        type: integer
      measurement:
        type: string
      quantity:
        type: number
      recipe_id:
        type: integer
      recipe_ingredient_id:
        type: integer
    type: object
  models.RecipeIngredientRequest:
    properties:
      ingredient_id:
        type: integer
      measurement:
        type: string
      quantity:
        type: number
      recipe_id:
        type: integer
    required:
    - ingredient_id
    - measurement
    - quantity
    - recipe_id
    type: object
...
  models.RecipeStep:
    properties:
      recipe_id:
        type: integer
      recipe_step_id:
        type: integer
      step_description:
        type: string
      step_id:
        type: integer
      step_number:
        type: integer
    type: object
  models.RecipeStepRequest:
    properties:
      recipe_id:
        type: integer
      recipe_step_id:
        type: integer
      step_description:
        type: string
      step_number:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /recipe-ingredients:
    get:
      consumes:
      - application/json
      description: Get all recipe ingredients from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RecipeIngredient'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all recipe ingredients
      tags:
      - recipe_ingredients
    post:
      consumes:
      - application/json
      description: Add a new recipe ingredient to the database
      parameters:
      - description: Add recipe ingredient
        in: body
        name: recipe_ingredient
        required: true
        schema:
          $ref: '#/definitions/models.RecipeIngredientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RecipeIngredient'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new recipe ingredient
      tags:
      - recipe_ingredients
  /recipe-ingredients/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a recipe ingredient by ID from the database
      parameters:
      - description: Recipe Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Recipe ingredient deleted
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recipe ingredient not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a recipe ingredient
      tags:
      - recipe_ingredients
    get:
      consumes:
      - application/json
      description: Get details of a specific recipe ingredient by ID
      parameters:
      - description: Recipe Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recipe ingredient retrieved
          schema:
            $ref: '#/definitions/models.RecipeIngredient'
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recipe ingredient not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a single recipe ingredient
      tags:
      - recipe_ingredients
    put:
      consumes:
      - application/json
      description: Update a recipe ingredient by ID
      parameters:
      - description: Recipe Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update recipe ingredient
        in: body
        name: recipe_ingredient
        required: true
        schema:
          $ref: '#/definitions/models.RecipeIngredient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecipeIngredient'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a recipe ingredient
      tags:
      - recipe_ingredients
  /recipe-steps:
    post:
      consumes:
      - application/json
      description: Add a new recipe step to the database
      parameters:
      - description: Add recipe step
        in: body
        name: recipe_step
        required: true
        schema:
          $ref: '#/definitions/models.RecipeStepRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RecipeStep'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new recipe step
      tags:
      - recipe_steps
  /recipe-steps/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a recipe step from the database
      parameters:
      - description: Recipe Step ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a recipe step
      tags:
      - recipe_steps
    get:
      consumes:
      - application/json
      description: Get a single recipe step from the database by ID
      parameters:
      - description: Recipe Step ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecipeStep'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a recipe step by ID
      tags:
      - recipe_steps
    put:
      consumes:
      - application/json
      description: Update a recipe step by ID
      parameters:
      - description: Recipe Step ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update recipe step
        in: body
        name: recipe_step
        required: true
        schema:
          $ref: '#/definitions/models.RecipeStep'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecipeStep'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a recipe step
      tags:
      - recipe_steps
...
swagger: "2.0"